backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
yt-dlp==2024.10.22
youtube-transcript-api==0.6.1
sentence-transformers==2.2.2
transformers==4.34.0
faiss-cpu==1.7.4
pydantic==1.10.11
firebase-admin==6.3.0
openai-whisper==20231117
ffmpeg-python==0.2.0
python-multipart==0.0.6
aiofiles==23.2.1

#pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu129